moves files inside homepage folder
