+++ components/homePage/index.js
@@ -1,0 +1,8 @@
+import About from "./About";
+import Internship from "./Internship";
+import MentorSection from "./MentorSection";
+import Service from "./Service";
+import TeamProfile from "./TeamProfile";
+import Testimonial from "./Testimonial";
+
+export { About, Internship, MentorSection, Service, TeamProfile, Testimonial };
